+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.1.0
-
 - Working, initial version
+
+## 0.1.1
+- Added --handleHoleDepth option, allowing you to just mark the hole and not drill through
+- Fixed a bug with single hole handles
+